api-types, ch-remote, fuzz, vmm: split `VmSendMigrationData`
Part of the move to an api-types crate.

On-behalf-of: SAP julian.schindel@sap.com
