+++ fuzz/fuzz_targets/http_api.rs
@@ -14,9 +14,9 @@ use libfuzzer_sys::{fuzz_target, Corpus};
 use micro_http::Request;
 use vm_migration::MigratableError;
 use vmm::api::http::*;
-use vmm::api::{ApiRequest, RequestHandler, VmInfoResponse, VmSendMigrationData};
+use vmm::api::{ApiRequest, RequestHandler, VmInfoResponse};
 use vmm::config::RestoreConfig;
-use vmm::migration::VmReceiveMigrationData;
+use vmm::migration::{VmReceiveMigrationData, VmSendMigrationData};
 use vmm::vm::{Error as VmError, VmState};
 use vmm::vm_config::*;
 use vmm::{EpollContext, EpollDispatch};
